+++ src/main/java/com/hongik/pcrc/allinone/auth/ui/controller/EmailController.java
@@ -3,10 +3,10 @@ package com.hongik.pcrc.allinone.auth.ui.controller;
 //import com.hongik.pcrc.allinone.application.service.EmailRedisService;
 import com.hongik.pcrc.allinone.auth.application.service.EmailService;
 import com.hongik.pcrc.allinone.auth.ui.requestBody.EmailVerifyRequest;
-import com.hongik.pcrc.allinone.auth.ui.view.ApiResponseView;
 import com.hongik.pcrc.allinone.auth.ui.view.Auth.SuccessView;
 import com.hongik.pcrc.allinone.exception.AllInOneException;
 import com.hongik.pcrc.allinone.exception.MessageType;
+import com.hongik.pcrc.allinone.exception.view.ApiResponseView;
 import org.springframework.http.ResponseEntity;
 import org.springframework.util.ObjectUtils;
 import org.springframework.web.bind.annotation.PostMapping;
