+++ src/main/java/com/hongik/pcrc/allinone/exception/MessageType.java
@@ -11,7 +11,7 @@ public enum MessageType {
     NOT_FOUND ("No data was found for the server. Please refer to parameter description.", HttpStatus.NOT_FOUND),
     INTERNAL_SERVER_ERROR ("An error occurred inside the server.", HttpStatus.INTERNAL_SERVER_ERROR),
     CONFLICT ("Already exists data.", HttpStatus.CONFLICT),
-    FORBIDDEN ("Authentication is denied.", HttpStatus.FORBIDDEN);
+    FORBIDDEN ("Authentication failed.", HttpStatus.FORBIDDEN);
 
     private final String message;
     private final HttpStatus status;
